fixed data type

+++ train/tasks/semantic/modules/trainer.py
@@ -651,7 +651,7 @@ class Trainer():
         proj_mask = proj_mask.cuda().long()
         proj_mask_t = proj_mask_t.cuda()
         proj_labels = proj_labels * proj_mask
-        in_vol = in_vol * proj_mask_t.unsqueeze(1) 
+        in_vol = in_vol * proj_mask_t.float().unsqueeze(1)
 
         ##############################
         # supervised learning (method3)
